+++ tests/test_squared_distance_to_set.py
@@ -34,7 +34,9 @@ class SquaredDistanceToSetTests(unittest.TestCase):
                 """
 fn dist_to_axis_codegen(
     x: &[{{ scalar_type }}],
+    w: &[{{ scalar_type }}],
 ) -> {{ scalar_type }} {
+    let _ = w;
     0.5_{{ scalar_type }} * x[1] * x[1]
 }
 """